compressdev: remove inline from enqueue/dequeue fns
Enqueue and dequeue functions were inline functions
in order to decrease the offload cost.
However, this requires some structures like
rte_compressdev and rte_compressdev_data to be public,
which they should not be.
Considering that the offload cost is going to be minimal
compared to the whole compression operation, it is
a better idea to make these functions non inline
(ethdev is actually going to take this approach in the near future).

Signed-off-by: Pablo de Lara <pablo.de.lara.guarch@intel.com>