compressdev: move internal structs into new file
Now that there is no need to expose internal structures,
these can be moved into a separate file, that will only be used
internally (in the library and also by the PMDs, but not by the applications).

This way, the only public API that will be used by apps is
in rte_compressdev.h.

Signed-off-by: Pablo de Lara <pablo.de.lara.guarch@intel.com>