compressdev: free device memory at device closure
When calling rte_compressdev_close(), the resources allocated
for the device at configuration (queue pairs, internal private data)
should be freed.

After calling this function, the device is in the same situation
as the one before calling rte_compressdev_configure(), so in order to
use it again, this latter function must be called, queue pairs must
be set up and devices must be restarted.

Signed-off-by: Pablo de Lara <pablo.de.lara.guarch@intel.com>